pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-autobot"
version = "0.1.4"
description = "Reusable Python toolkit for LinkedIn lead collection, outreach automation, and lead analytics."
readme = "PYPI_README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
  { name = "Ayoub Ardem" }
]
keywords = ["linkedin", "automation", "lead-generation", "playwright", "analytics"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Internet :: WWW/HTTP :: Browsers",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = []

[project.optional-dependencies]
automation = [
  "playwright>=1.50.0",
]
analytics = [
  "langchain>=0.3.0",
  "langchain-google-genai>=2.1.0",
]
django = [
  "django>=5.0",
  "djangorestframework>=3.15.0",
]
dev = [
  "build>=1.2.0",
  "pytest>=8.0.0",
  "twine>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/AyoubArdem/LinkedinAutomationBot"
Documentation = "https://github.com/AyoubArdem/LinkedinAutomationBot"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
linkedin_autobot = ["py.typed"]

[tool.setuptools.packages.find]
where = ["src"]
